+++ src/lib/helpers.ts
@@ -18,7 +18,7 @@ export const BASE_URI =
 
 // Same files served straight from S3: no rate limit, but uncached and slower.
 // Fallback once the daily request limit of the primary endpoint is hit.
-export const SLOW_BASE_URI = 'https://openmeteo.s3.amazonaws.com/data_spatial';
+export const S3_BASE_URI = 'https://openmeteo.s3.amazonaws.com/data_spatial';
 
 export const hashValue = (val: string): string => {
 	// FNV-1a 32-bit – synchronous, fast, and sufficient for cache-busting keys.
